+++ include/MyECS/RandomAccessor.h
@@ -1,0 +1,16 @@
+//
+// Created by Admin on 28/03/2025.
+//
+
+#pragma once
+ 
+ #include "CmptType.h"
+ 
+ #include <set>
+ 
+ namespace My::MyECS {
+ class RandomAccessor {
+ public:
+   CmptAccessTypeSet types;
+ };
+ }