+++ include/MyECS/World.h
@@ -35,7 +35,7 @@ class World {
   // 4. run commands in main thread
   void Update();
 
-  void AddCommand(std::function<void()> command, size_t layer = 0);
+  void AddCommand(std::function<void()> command, int layer = 0);
 
   // after running Update()
   // you can use graphviz to vistualize the graph
@@ -108,7 +108,7 @@ class World {
   Pool<Job> jobPool;
 
   // command
-  std::map<size_t, std::vector<std::function<void()>>> commandBuffer;
+  std::map<int, std::vector<std::function<void()>>> commandBuffer;
   std::mutex commandBufferMutex;
   void RunCommands();